requirements.txt
streamlit==1.37.0
pandas==2.2.2
bcrypt==4.2.0
streamlit-option-menu==0.3.12
gspread==6.1.2
google-auth==2.34.0
python-dateutil==2.8.2
reportlab==4.0.9
openpyxl==3.1.5